updated add/remove
