+++ src/components/MovieList/MovieList.js
@@ -1,10 +1,6 @@
 import React, { Component } from 'react';
-import { firebaseUrl } from '../../../config';
 import Movie from '../Movie';
 import './MovieList.scss';
-import Rebase from 're-base';
-
-const baseUrl = Rebase.createClass(firebaseUrl);
 
 
 export default class MovieList extends Component {
@@ -12,34 +8,18 @@ export default class MovieList extends Component {
     super(props);
 
     this.state = {
-      moviesToWatch: [],
       searchString: ''
     };
 
     this.handleFilter = this.handleFilter.bind(this);
-    this.remove = this.remove.bind(this);
-  }
-
-  componentDidMount() {
-    baseUrl.syncState('movieList', {
-      context: this,
-      state: 'moviesToWatch',
-      asArray: true
-    });
   }
 
   handleFilter(event) {
     this.setState({ searchString: event.target.value.toLowerCase() });
   }
 
-  remove(index) {
-    let newList = this.state.moviesToWatch;
-    newList.splice(index, 1);
-    this.setState({ moviesToWatch: newList });
-  }
-
   render() {
-    let listOfMovies = this.state.moviesToWatch,
+    let listOfMovies = this.props.moviesToWatch,
         searchString = this.state.searchString;
 
     if (searchString.length) {
@@ -47,11 +27,12 @@ export default class MovieList extends Component {
     }
 
     if (listOfMovies.length) {
-      listOfMovies = listOfMovies.map((movie, i) => <Movie key={i} index={i} remove={this.remove} movies={movie} />);
+      listOfMovies = listOfMovies.map((movie, i) => <Movie key={i} index={i} remove={this.props.removeMovie} movies={movie} />);
     }
+
     return (
       <div className="movie-list-wrapper">
-        <input type="text" placeholder="Search saved movies" onChange={this.handleFilter} />
+        <input type="text" className="filter" placeholder="Search saved movies" onChange={this.handleFilter} />
         <ul className="movie-list-container">{listOfMovies}</ul>
       </div>
     );
