+++ src/components/MovieList/MovieList.scss
@@ -1,6 +1,18 @@
 .movie-list-wrapper {
   float: left;
   width: 100%;
+
+  .filter {
+    width: 100%;
+    height: 40px;
+    max-width: 400px;
+    font-size: 2rem;
+    padding-right: 10px;
+    padding-left: 10px;
+    margin-right: 5px;
+    margin-left: 5px;
+    margin-bottom: 20px;
+  }
 }
 
 .movie-list-container {
