+++ src/components/Overlay/Overlay.scss
@@ -7,7 +7,7 @@
   opacity: 1;
   height: 100%;
   width: 100%;
-  background-color: rgba(255, 255, 255, 0.95);
+  background-color: rgba(255, 255, 255, 0.97);
   overflow-y: scroll;
   overflow-x: hidden;
   z-index: 2;
