+++ src/components/PosterNotFound/PosterNotFound.scss
@@ -7,8 +7,8 @@
 
 	&-info {
 		position: absolute;
-    padding-left: 5px;
-    padding-right: 5px;
+    padding-left: 10px;
+    padding-right: 10px;
     left: 0;
     top: 50%;
     width: 100%;
@@ -23,5 +23,6 @@
 		display: block;
 		text-align: center;
 		font-size: 5rem;
+    margin-top: 20px;
 	}
 }
